Update ocean filtering plugins
